Use fully-qualified names in macro quasiquotes
Use fully-qualified names in macro quasiquotes to
ensure that the correct type is always used and that
API users never need to import any types for the
macros to work.

Fixes a bug where users had to import
`com.github.andyglow.jsonschema.TypeSignature`
in order to use `json.Json.schema`.